chore: organize gfx model (#8766)
### Changed

- Reorganize the gfx model
