+++ packages/framework/block-std/src/gfx/controller.ts
@@ -11,7 +11,9 @@ import {
 
 import type { BlockStdScope } from '../scope/block-std-scope.js';
 import type { BlockComponent } from '../view/index.js';
-import type { SurfaceBlockModel } from './surface/surface-model.js';
+import type { PointTestOptions } from './model/base.js';
+import type { GfxModel } from './model/model.js';
+import type { SurfaceBlockModel } from './model/surface/surface-model.js';
 
 import { LifeCycleWatcher } from '../extension/lifecycle-watcher.js';
 import { onSurfaceAdded } from '../utils/gfx.js';
@@ -19,16 +21,15 @@ import {
   GfxClassExtenderIdentifier,
   GfxExtensionIdentifier,
 } from './extension.js';
-import { GfxBlockElementModel, type GfxModel } from './gfx-block-model.js';
 import { GridManager } from './grid.js';
 import { gfxControllerKey } from './identifiers.js';
 import { KeyboardController } from './keyboard.js';
 import { LayerManager } from './layer.js';
+import { GfxBlockElementModel } from './model/gfx-block-model.js';
 import {
   GfxGroupLikeElementModel,
   GfxPrimitiveElementModel,
-  type PointTestOptions,
-} from './surface/element-model.js';
+} from './model/surface/element-model.js';
 import { Viewport } from './viewport.js';
 
 export class GfxController extends LifeCycleWatcher {
