+++ packages/framework/block-std/src/gfx/model/model.ts
@@ -1,0 +1,4 @@
+import type { GfxBlockElementModel } from './gfx-block-model.js';
+import type { GfxPrimitiveElementModel } from './surface/element-model.js';
+
+export type GfxModel = GfxBlockElementModel | GfxPrimitiveElementModel;
