+++ packages/framework/block-std/src/gfx/model/surface/container-element.ts
@@ -1,14 +1,24 @@
-import type { GfxModel } from '../gfx-block-model.js';
-import type { GfxCompatibleProps } from './element-model.js';
+import type { GfxCompatibleInterface } from '../base.js';
+import type { GfxModel } from '../model.js';
 
+/**
+ * The symbol to mark a model as a container.
+ */
 export const gfxContainerSymbol = Symbol('GfxContainerElement');
 
+/**
+ * Check if the element is a container element.
+ */
 export const isGfxContainerElm = (elm: unknown): elm is GfxContainerElement => {
   if (typeof elm !== 'object' || elm === null) return false;
   return gfxContainerSymbol in elm && elm[gfxContainerSymbol] === true;
 };
 
-export interface GfxContainerElement extends GfxCompatibleProps {
+/**
+ * GfxContainerElement is a model that can contain other models.
+ * It just like a group that in common graphic software.
+ */
+export interface GfxContainerElement extends GfxCompatibleInterface {
   [gfxContainerSymbol]: true;
   childIds: string[];
