+++ packages/framework/block-std/src/gfx/model/surface/decorators/common.ts
