+++ packages/framework/block-std/src/gfx/model/surface/decorators/derive.ts
