+++ packages/framework/block-std/src/utils/layer.ts
@@ -2,14 +2,15 @@ import type { Doc } from '@blocksuite/store';
 
 import { nToLast } from '@blocksuite/global/utils';
 
-import type { GfxBlockElementModel, GfxModel } from '../gfx/gfx-block-model.js';
 import type { Layer } from '../gfx/layer.js';
-import type { SurfaceBlockModel } from '../gfx/surface/surface-model.js';
+import type { GfxBlockElementModel } from '../gfx/model/gfx-block-model.js';
+import type { GfxModel } from '../gfx/model/model.js';
+import type { SurfaceBlockModel } from '../gfx/model/surface/surface-model.js';
 
 import {
   type GfxContainerElement,
   isGfxContainerElm,
-} from '../gfx/surface/container-element.js';
+} from '../gfx/model/surface/container-element.js';
 
 export function getLayerEndZIndex(layers: Layer[], layerIndex: number) {
   const layer = layers[layerIndex];
